solve circular reference between `JwtTokenManager` and `NacosAuthConfig`
